+++ CHANGELOG.md
@@ -3,6 +3,7 @@ next-version
 
 - cli: add support for writing json --report /dev/stdout.
 - iterator: improve end of line detection in ansible one-line output.
+- model: add support for golang default timestamps.
 
 0.15.0
 ======